python/wandb_sweep_example/requirements.freeze.txt
certifi==2022.5.18.1
charset-normalizer==2.0.12
click==8.1.3
docker-pycreds==0.4.0
gitdb==4.0.9
GitPython==3.1.27
idna==3.3
numpy==1.22.4
pathtools==0.1.2
pkg_resources==0.0.0
promise==2.3
protobuf==3.20.1
psutil==5.9.1
PyYAML==6.0
requests==2.28.0
sentry-sdk==1.5.12
setproctitle==1.2.3
shortuuid==1.0.9
six==1.16.0
smmap==5.0.0
urllib3==1.26.9
wandb==0.12.18
